+++ diary1801204/src/diary/servlet/DutyUpdateServlet.java
@@ -2,6 +2,8 @@ package diary.servlet;
 
 import diary.bean.DiaryBeans;
 import diary.bean.DutyBeans;
+import diary.bean.TeacherBeans;
+import diary.commmon.TeacherErrorCheck;
 import diary.dao.DutyDao;
 import diary.dao.StudentDiaryDao;
 
@@ -30,20 +32,30 @@ public class DutyUpdateServlet extends HttpServlet {
         System.out.println("DutyUpdateServlet");
         ///////////////////////////////////////////////////////////////////////////////////////////////////
 
+        //ログイン済みかチェックする
         HttpSession session = request.getSession();
-        DutyBeans duty_beans = (DutyBeans) session.getAttribute("duty-beans");
+        TeacherBeans teacher_beans = (TeacherBeans) session.getAttribute("teacher-beans");
 
-        DutyDao duty_dao = new DutyDao();
-        duty_dao.updateDutyToDb(duty_beans);
+        TeacherErrorCheck error_check = new TeacherErrorCheck();
+        boolean is_login = error_check.checkLogin(teacher_beans);
 
-        DiaryBeans diary_beans = new DiaryBeans();
-        diary_beans.setClass_code(duty_beans.getClass_code());
-        diary_beans.setInsert_date(duty_beans.getInsert_date());
+        if (is_login) {
+            DutyBeans duty_beans = (DutyBeans) session.getAttribute("duty-beans");
 
-        StudentDiaryDao diary_dao = new StudentDiaryDao();
-        diary_dao.deleteDiaryFromDb(diary_beans);
+            DutyDao duty_dao = new DutyDao();
+            duty_dao.updateDutyToDb(duty_beans);
 
-        session.removeAttribute("duty-beans");
-        response.sendRedirect("dutyupdatecomplete");
+            DiaryBeans diary_beans = new DiaryBeans();
+            diary_beans.setClass_code(duty_beans.getClass_code());
+            diary_beans.setInsert_date(duty_beans.getInsert_date());
+
+            StudentDiaryDao diary_dao = new StudentDiaryDao();
+            diary_dao.deleteDiaryFromDb(diary_beans);
+
+            session.removeAttribute("duty-beans");
+            response.sendRedirect("dutyupdatecomplete");
+        } else {
+            response.sendRedirect("teachererror");
+        }
     }
 }
